+++ push_notifications/api/rest_framework.py
@@ -1,5 +1,8 @@
+<<<<<<< HEAD
 from __future__ import absolute_import
 
+=======
+>>>>>>> upstream/master
 from rest_framework import permissions, status
 from rest_framework.fields import IntegerField
 from rest_framework.response import Response
@@ -140,9 +143,15 @@ class DeviceViewSetMixin(object):
 	def create(self, request, *args, **kwargs):
 		serializer = None
 		is_update = False
+<<<<<<< HEAD
 		if SETTINGS.get("UPDATE_ON_DUPLICATE_REG_ID") and "registration_id" in request.data:
 			instance = self.queryset.model.objects.filter(
 				registration_id=request.data["registration_id"]
+=======
+		if SETTINGS.get("UPDATE_ON_DUPLICATE_REG_ID") and self.lookup_field in request.data:
+			instance = self.queryset.model.objects.filter(
+				registration_id=request.data[self.lookup_field]
+>>>>>>> upstream/master
 			).first()
 			if instance:
 				serializer = self.get_serializer(instance, data=request.data)