+++ push_notifications/exceptions.py
@@ -1,3 +1,6 @@
+<<<<<<< HEAD
 
+=======
+>>>>>>> upstream/master
 class NotificationError(Exception):
 	pass