+++ push_notifications/fields.py
@@ -4,7 +4,10 @@ import struct
 from django import forms
 from django.core.validators import MaxValueValidator, MinValueValidator, RegexValidator
 from django.db import connection, models
+<<<<<<< HEAD
 from django.utils import six
+=======
+>>>>>>> upstream/master
 from django.utils.translation import ugettext_lazy as _
 
 
@@ -55,7 +58,7 @@ class HexadecimalField(forms.CharField):
 
 	def prepare_value(self, value):
 		# converts bigint from db to hex before it is displayed in admin
-		if value and not isinstance(value, six.string_types) \
+		if value and not isinstance(value, str) \
 			and connection.vendor in ("mysql", "sqlite"):
 			value = _unsigned_integer_to_hex_string(value)
 		return super(forms.CharField, self).prepare_value(value)
@@ -92,13 +95,13 @@ class HexIntegerField(models.BigIntegerField):
 		""" Return the integer value to be stored from the hex string """
 		if value is None or value == "":
 			return None
-		if isinstance(value, six.string_types):
+		if isinstance(value, str):
 			value = _hex_string_to_unsigned_integer(value)
 		if _using_signed_storage():
 			value = _unsigned_to_signed_integer(value)
 		return value
 
-	def from_db_value(self, value, expression, connection, context):
+	def from_db_value(self, value, *args):
 		""" Return an unsigned int representation from all db backends """
 		if value is None:
 			return value
@@ -108,7 +111,7 @@ class HexIntegerField(models.BigIntegerField):
 
 	def to_python(self, value):
 		""" Return a str representation of the hexadecimal """
-		if isinstance(value, six.string_types):
+		if isinstance(value, str):
 			return value
 		if value is None:
 			return value