+++ push_notifications/models.py
@@ -1,7 +1,9 @@
+<<<<<<< HEAD
 from __future__ import unicode_literals
 
+=======
+>>>>>>> upstream/master
 from django.db import models
-from django.utils.encoding import python_2_unicode_compatible
 from django.utils.translation import ugettext_lazy as _
 
 from .fields import HexIntegerField
@@ -20,7 +22,6 @@ BROWSER_TYPES = (
 )
 
 
-@python_2_unicode_compatible
 class Device(models.Model):
 	name = models.CharField(max_length=255, verbose_name=_("Name"), blank=True, null=True)
 	active = models.BooleanField(
@@ -47,9 +48,15 @@ class Device(models.Model):
 
 	def __str__(self):
 		return (
+<<<<<<< HEAD
 			self.name
 			or str(self.device_id or "")
 			or "%s for %s" % (self.__class__.__name__, self.user or "unknown user")
+=======
+			self.name or
+			str(self.device_id or "") or
+			"{} for {}".format(self.__class__.__name__, self.user or "unknown user")
+>>>>>>> upstream/master
 		)
 
 
@@ -94,10 +101,17 @@ class GCMDevice(Device):
 		verbose_name=_("Device ID"), blank=True, null=True, db_index=True,
 		help_text=_("ANDROID_ID / TelephonyManager.getDeviceId() (always as hex)")
 	)
+<<<<<<< HEAD
 	registration_id = models.TextField(verbose_name=_("Registration ID"))
 	cloud_message_type = models.CharField(
 		verbose_name=_("Cloud Message Type"), max_length=3,
 		choices=CLOUD_MESSAGE_TYPES, default="FCM",
+=======
+	registration_id = models.TextField(verbose_name=_("Registration ID"), unique=SETTINGS["UNIQUE_REG_ID"])
+	cloud_message_type = models.CharField(
+		verbose_name=_("Cloud Message Type"), max_length=3,
+		choices=CLOUD_MESSAGE_TYPES, default="GCM",
+>>>>>>> upstream/master
 		help_text=_("You should choose FCM or GCM")
 	)
 	objects = GCMDeviceManager()
@@ -124,7 +138,11 @@ class APNSDeviceManager(models.Manager):
 
 
 class APNSDeviceQuerySet(models.query.QuerySet):
+<<<<<<< HEAD
 	def send_message(self, message, certfile=None, **kwargs):
+=======
+	def send_message(self, message, creds=None, **kwargs):
+>>>>>>> upstream/master
 		if self:
 			from .apns import apns_send_bulk_message
 
@@ -137,7 +155,11 @@ class APNSDeviceQuerySet(models.query.QuerySet):
 				)
 				r = apns_send_bulk_message(
 					registration_ids=reg_ids, alert=message, application_id=app_id,
+<<<<<<< HEAD
 					certfile=certfile, **kwargs
+=======
+					creds=creds, **kwargs
+>>>>>>> upstream/master
 				)
 				if hasattr(r, "keys"):
 					res += [r]
@@ -152,7 +174,11 @@ class APNSDevice(Device):
 		help_text="UDID / UIDevice.identifierForVendor()"
 	)
 	registration_id = models.CharField(
+<<<<<<< HEAD
 		verbose_name=_("Registration ID"), max_length=200, unique=True
+=======
+		verbose_name=_("Registration ID"), max_length=200, unique=SETTINGS["UNIQUE_REG_ID"]
+>>>>>>> upstream/master
 	)
 
 	objects = APNSDeviceManager()
@@ -160,13 +186,21 @@ class APNSDevice(Device):
 	class Meta:
 		verbose_name = _("APNS device")
 
+<<<<<<< HEAD
 	def send_message(self, message, certfile=None, **kwargs):
+=======
+	def send_message(self, message, creds=None, **kwargs):
+>>>>>>> upstream/master
 		from .apns import apns_send_message
 
 		return apns_send_message(
 			registration_id=self.registration_id,
 			alert=message,
+<<<<<<< HEAD
 			application_id=self.application_id, certfile=certfile,
+=======
+			application_id=self.application_id, creds=creds,
+>>>>>>> upstream/master
 			**kwargs
 		)
 
@@ -196,6 +230,7 @@ class WNSDeviceQuerySet(models.query.QuerySet):
 
 		return res
 
+<<<<<<< HEAD
 
 class WNSDevice(Device):
 	device_id = models.UUIDField(
@@ -204,6 +239,16 @@ class WNSDevice(Device):
 	)
 	registration_id = models.TextField(verbose_name=_("Notification URI"))
 
+=======
+
+class WNSDevice(Device):
+	device_id = models.UUIDField(
+		verbose_name=_("Device ID"), blank=True, null=True, db_index=True,
+		help_text=_("GUID()")
+	)
+	registration_id = models.TextField(verbose_name=_("Notification URI"), unique=SETTINGS["UNIQUE_REG_ID"])
+
+>>>>>>> upstream/master
 	objects = WNSDeviceManager()
 
 	class Meta:
@@ -234,7 +279,11 @@ class WebPushDeviceQuerySet(models.query.QuerySet):
 
 
 class WebPushDevice(Device):
+<<<<<<< HEAD
 	registration_id = models.TextField(verbose_name=_("Registration ID"))
+=======
+	registration_id = models.TextField(verbose_name=_("Registration ID"), unique=SETTINGS["UNIQUE_REG_ID"])
+>>>>>>> upstream/master
 	p256dh = models.CharField(
 		verbose_name=_("User public encryption key"),
 		max_length=88)