+++ tests/test_app_config.py
@@ -128,10 +128,24 @@ class AppConfigTestCase(TestCase):
 			}
 		}
 
+<<<<<<< HEAD
 		with self.assertRaises(ImproperlyConfigured):
 			AppConfig(PUSH_SETTINGS)
 
 		# all optional settings have default values
+=======
+		with self.assertRaises(ImproperlyConfigured) as ic:
+			AppConfig(PUSH_SETTINGS)
+
+		self.assertEqual(
+			str(ic.exception),
+			("PUSH_NOTIFICATIONS_SETTINGS.APPLICATIONS[\"my_apns_app\"][\"('CERTIFICATE', ['AUTH_KEY_PATH', 'AUTH_KEY_ID', 'TEAM_ID'])\"] is missing.")
+		)
+
+		#
+		# certificate settings, with optional settings having default values
+		#
+>>>>>>> upstream/master
 		PUSH_SETTINGS = {
 			"APPLICATIONS": {
 				"my_apns_app": {
@@ -147,6 +161,28 @@ class AppConfigTestCase(TestCase):
 		assert app_config["USE_SANDBOX"] is False
 		assert app_config["USE_ALTERNATIVE_PORT"] is False
 
+<<<<<<< HEAD
+=======
+		# certificate settings, with optional settings having default values
+		#
+		PUSH_SETTINGS = {
+			"APPLICATIONS": {
+				"my_apns_app": {
+					"PLATFORM": "APNS",
+					"AUTH_KEY_PATH": path,
+					"AUTH_KEY_ID": "123456",
+					"TEAM_ID": "123456",
+				}
+			}
+		}
+
+		manager = AppConfig(PUSH_SETTINGS)
+		app_config = manager._settings["APPLICATIONS"]["my_apns_app"]
+
+		assert app_config["USE_SANDBOX"] is False
+		assert app_config["USE_ALTERNATIVE_PORT"] is False
+
+>>>>>>> upstream/master
 	def test_get_allowed_settings_fcm(self):
 		"""Verify the settings allowed for FCM platform."""