+++ tests/test_models.py
@@ -95,7 +95,11 @@ class GCMModelTestCase(TestCase):
 				GCMDevice.objects.all().send_message(
 					None, extra={"title": "bar", "body": "foo"}, collapse_key="test_key"
 				)
+<<<<<<< HEAD
 				self.assertEquals(p.call_count, 2)
+=======
+				self.assertEqual(p.call_count, 2)
+>>>>>>> upstream/master
 				p.assert_has_calls([
 					mock.call(
 						json.dumps({