+++ tests/test_wns.py
@@ -88,12 +88,21 @@ class WNSDictToXmlSchemaTestCase(TestCase):
 		xml_tree = dict_to_xml_schema(xml_data)
 		self.assertEqual(xml_tree.tag, "toast")
 		self.assertEqual(xml_tree.attrib, {"key": "value"})
+<<<<<<< HEAD
 		visual = xml_tree.getchildren()[0]
 		self.assertEqual(visual.tag, "visual")
 		binding = visual.getchildren()[0]
 		self.assertEqual(binding.tag, "binding")
 		self.assertEqual(binding.attrib, {"template": "ToastText01"})
 		text = binding.getchildren()[0]
+=======
+		visual = list(xml_tree)[0]
+		self.assertEqual(visual.tag, "visual")
+		binding = list(visual)[0]
+		self.assertEqual(binding.tag, "binding")
+		self.assertEqual(binding.attrib, {"template": "ToastText01"})
+		text = list(binding)[0]
+>>>>>>> upstream/master
 		self.assertEqual(text.tag, "text")
 		self.assertEqual(text.attrib, {"id": "1"})
 		self.assertEqual(text.text, "toast notification")
@@ -125,6 +134,7 @@ class WNSDictToXmlSchemaTestCase(TestCase):
 		xml_tree = dict_to_xml_schema(xml_data)
 		self.assertEqual(xml_tree.tag, "toast")
 		self.assertEqual(xml_tree.attrib, {"key": "value"})
+<<<<<<< HEAD
 		visual = xml_tree.getchildren()[0]
 		self.assertEqual(visual.tag, "visual")
 		binding = visual.getchildren()[0]
@@ -132,6 +142,14 @@ class WNSDictToXmlSchemaTestCase(TestCase):
 		self.assertEqual(binding.attrib, {"template": "ToastText02"})
 		children = binding.getchildren()
 		self.assertEqual(len(children), 2)
+=======
+		visual = list(xml_tree)[0]
+		self.assertEqual(visual.tag, "visual")
+		binding = list(visual)[0]
+		self.assertEqual(binding.tag, "binding")
+		self.assertEqual(binding.attrib, {"template": "ToastText02"})
+		self.assertEqual(len(list(binding)), 2)
+>>>>>>> upstream/master
 
 	def test_create_two_multi_sub_element_xml_from_dict(self):
 		xml_data = {
@@ -166,6 +184,7 @@ class WNSDictToXmlSchemaTestCase(TestCase):
 		xml_tree = dict_to_xml_schema(xml_data)
 		self.assertEqual(xml_tree.tag, "toast")
 		self.assertEqual(xml_tree.attrib, {"key": "value"})
+<<<<<<< HEAD
 		visual = xml_tree.getchildren()[0]
 		self.assertEqual(visual.tag, "visual")
 		binding = visual.getchildren()[0]
@@ -173,3 +192,11 @@ class WNSDictToXmlSchemaTestCase(TestCase):
 		self.assertEqual(binding.attrib, {"template": "ToastText02"})
 		children = binding.getchildren()
 		self.assertEqual(len(children), 4)
+=======
+		visual = list(xml_tree)[0]
+		self.assertEqual(visual.tag, "visual")
+		binding = list(visual)[0]
+		self.assertEqual(binding.tag, "binding")
+		self.assertEqual(binding.attrib, {"template": "ToastText02"})
+		self.assertEqual(len(list(binding)), 4)
+>>>>>>> upstream/master